Convert to fixed point math for speedup

+++ hdma.h
@@ -22,6 +22,6 @@ void HdmaEngine_SetupPaletteHdma(void);
 void HdmaEngine_SetupBgScrollHdma(void);
 void HdmaEngine_UpdateBgScrollValues();
 
-void HdmaEngine_GeneratePaletteTable(uint16_t * table_ptr, uint16_t pal_start, uint16_t entries, uint16_t target_color, float alpha, int16_t height);
+void HdmaEngine_GeneratePaletteTable(uint16_t * table_ptr, uint16_t pal_start, uint16_t entries, uint16_t target_color, uint16_t alpha, uint16_t height);
 
 void HdmaEngine_EnableHdma(void);
